fix(ollama): move OllamaModels back to Ollama client module
