+++ agents/agents-core/src/commonTest/kotlin/ai/koog/agents/core/feature/choice/ChoiceSelectionStrategyTest.kt
@@ -7,7 +7,7 @@ import ai.koog.prompt.dsl.prompt
 import ai.koog.prompt.executor.model.LLMChoice
 import ai.koog.prompt.executor.model.PromptExecutor
 import ai.koog.prompt.llm.LLModel
-import ai.koog.prompt.llm.OllamaModels
+import ai.koog.prompt.executor.clients.ollama.OllamaModels
 import ai.koog.prompt.message.Message
 import ai.koog.prompt.message.ResponseMetaInfo
 import kotlinx.coroutines.flow.Flow
