+++ examples/src/main/kotlin/ai/koog/agents/example/calculator/OllamaCalculatorExample.kt
@@ -10,9 +10,9 @@ import ai.koog.agents.ext.tool.AskUser
 import ai.koog.agents.ext.tool.SayToUser
 import ai.koog.agents.features.eventHandler.feature.handleEvents
 import ai.koog.prompt.dsl.prompt
+import ai.koog.prompt.executor.clients.ollama.OllamaModels
 import ai.koog.prompt.executor.llms.all.simpleOllamaAIExecutor
 import ai.koog.prompt.executor.model.PromptExecutor
-import ai.koog.prompt.llm.OllamaModels
 import ai.koog.prompt.params.LLMParams
 import kotlinx.coroutines.runBlocking
 import kotlin.uuid.ExperimentalUuidApi
